pyproject.toml
[project]
name = "gaperon-finance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "huggingface-hub>=0.35.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pymupdf>=1.26.5",
    "slurmtui>=0.4.0",
    "vllm>=0.11.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.3",
]
